+++ rebase.sh
@@ -123,7 +123,10 @@ apply_code_package_changes() {
   
   # now apply again the changes
   override_json_file code/package.json
-  
+
+  # apply the replace
+  apply_replace code/package.json
+
   # resolve the change
   git add code/package.json > /dev/null 2>&1
 }
@@ -383,6 +386,26 @@ apply_changes() {
   git add "$filePath" > /dev/null 2>&1
 }
 
+# Apply changes for the given file, using multi line replace
+apply_changes_multi_line() {
+  local filePath="$1"
+
+  if [ -z "$filePath" ]; then
+     echo "Can not apply changes - the path was not passed"
+     exit 1;
+  fi
+
+  echo "  ⚙️ reworking $filePath..."
+  # reset the file from what is upstream
+  git checkout --theirs "$filePath" > /dev/null 2>&1
+
+  # now apply again the changes
+  apply_multi_line_replace "$filePath"
+
+  # resolve the change
+  git add "$filePath" > /dev/null 2>&1
+}
+
 # Will try to identify the conflicting files and for some of them it's easy to re-apply changes
 resolve_conflicts() {
   echo "⚠️  There are conflicting files, trying to solve..."
@@ -409,6 +432,8 @@ resolve_conflicts() {
     elif [[ "$conflictingFile" == "code/extensions/microsoft-authentication/package-lock.json" ]]; then
       apply_code_extensions_microsoft_authentication_package_lock_changes
     elif [[ "$conflictingFile" == "code/test/mcp/package.json" ]]; then
+      apply_package_changes_by_path "$conflictingFile"
+    elif [[ "$conflictingFile" == "code/test/monaco/package.json" ]]; then
       apply_package_changes_by_path "$conflictingFile"
     elif [[ "$conflictingFile" == "code/build/lib/mangle/index.ts" ]]; then
       apply_mangle_index_ts_changes
@@ -438,7 +463,11 @@ resolve_conflicts() {
       apply_changes "$conflictingFile"
     elif [[ "$conflictingFile" == "code/src/vs/server/node/serverEnvironmentService.ts" ]]; then
       apply_changes "$conflictingFile"
+    elif [[ "$conflictingFile" == "code/src/vs/platform/shell/node/shellEnv.ts" ]]; then
+      apply_changes "$conflictingFile"
     elif [[ "$conflictingFile" == "code/src/vs/server/node/extensionHostConnection.ts" ]]; then
+      apply_changes "$conflictingFile"
+    elif [[ "$conflictingFile" == "code/src/vs/server/node/remoteTerminalChannel.ts" ]]; then
       apply_changes "$conflictingFile"
     elif [[ "$conflictingFile" == "code/src/vs/code/browser/workbench/workbench.html" ]]; then
       apply_changes "$conflictingFile"
@@ -464,6 +493,10 @@ resolve_conflicts() {
       apply_multi_line_replace "$conflictingFile"
     elif [[ "$conflictingFile" == "code/src/vs/workbench/contrib/extensions/browser/extensionsWorkbenchService.ts" ]]; then
       apply_multi_line_replace "$conflictingFile"
+    elif [[ "$conflictingFile" == "code/build/gulpfile.cli.js" ]]; then
+      apply_changes_multi_line "$conflictingFile"
+    elif [[ "$conflictingFile" == "code/build/gulpfile.reh.js" ]]; then
+      apply_changes "$conflictingFile"
     else
       echo "$conflictingFile file cannot be automatically rebased. Aborting"
       exit 1